# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Ba7ath Forensic API
  version: 3.0.1
  description: >
    API d'authentification et endpoints métiers Ba7ath Forensic (JWT stateless, rate limiting, cache).
servers:
  - url: /api
    description: Local (préfixe API)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    X-RateLimit-Limit:
      description: Nombre max de requêtes autorisées sur la fenêtre en cours
      schema: { type: integer, example: 100 }
    X-RateLimit-Remaining:
      description: Requêtes restantes dans la fenêtre en cours
      schema: { type: integer, example: 87 }
    X-RateLimit-Reset:
      description: Timestamp Unix du reset de la fenêtre
      schema: { type: integer, example: 1694102400 }
    X-RateLimit-Window:
      description: Taille de la fenêtre en secondes
      schema: { type: integer, example: 900 }
    X-RateLimit-Action:
      description: Code logique du limiter appliqué
      schema: { type: string, example: login_attempt }
    RateLimit-Policy:
      description: Politique de limitation "<limit>;w=<window>"
      schema: { type: string, example: "20;w=3600" }
    RateLimit-Limit:
      description: Nombre max de requêtes autorisées dans la fenêtre
      schema: { type: integer, example: 20 }
    RateLimit-Remaining:
      description: Requêtes restantes dans la fenêtre courante
      schema: { type: integer, example: 19 }
    RateLimit-Reset:
      description: Secondes restantes avant reset de la fenêtre
      schema: { type: integer, example: 3540 }
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        error: { type: string, example: "Erreur lors de la connexion" }
        type: { type: string, example: "LOGIN_ERROR" }
        requestId: { type: string, example: "a1b2c3d4e5f6" }
...
    User:
      type: object
      properties:
        id: { type: string, example: "68be60ce3bb0afd3ce86b8f5" }
        email: { type: string, format: email }
        roles: { type: array, items: { type: string } }
        profile:
          type: object
          properties:
            firstName: { type: string }
            lastName: { type: string }
        security:
          type: object
          properties:
            lastLogin: { type: string, format: date-time }
            lastLoginIP: { type: string }
            lastActivity: { type: string, format: date-time }
            lastActivityIP: { type: string }
            lastActivityUA: { type: string }
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Connexion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                rememberMe: { type: boolean, default: false }
      responses:
        '200':
          description: Connexion réussie
          headers:
            X-RateLimit-Limit: { $ref: '#/components/headers/X-RateLimit-Limit' }
            X-RateLimit-Remaining: { $ref: '#/components/headers/X-RateLimit-Remaining' }
            X-RateLimit-Reset: { $ref: '#/components/headers/X-RateLimit-Reset' }
            X-RateLimit-Window: { $ref: '#/components/headers/X-RateLimit-Window' }
            X-RateLimit-Action: { $ref: '#/components/headers/X-RateLimit-Action' }
          content:
            application/json:
              schema:
                type: object
                required: [success, message, user, tokens, session]
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: "Connexion réussie" }
                  user: { $ref: '#/components/schemas/User' }
                  tokens:
                    type: object
                    properties:
                      accessToken: { type: string }
                      refreshToken: { type: string }
                  session:
                    type: object
                    properties:
                      id: { type: string }
                      expiresAt: { type: string, format: date-time }
                      deviceInfo: { type: object }
        '400':
          description: Données invalides
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Identifiants invalides
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Compte suspendu/non vérifié
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '423':
          description: Compte verrouillé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Trop de tentatives
          headers:
            Retry-After:
              description: Secondes avant de réessayer
              schema: { type: integer, example: 900 }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Erreur serveur interne
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /images/upload:
    post:
      tags: [images]
      summary: Upload image unique avec analyse rapide
      description: Upload d'un fichier image via multipart/form-data (champ "image")
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Image uploadée
          headers:
            RateLimit-Policy: { $ref: '#/components/headers/RateLimit-Policy' }
            RateLimit-Limit: { $ref: '#/components/headers/RateLimit-Limit' }
            RateLimit-Remaining: { $ref: '#/components/headers/RateLimit-Remaining' }
            RateLimit-Reset: { $ref: '#/components/headers/RateLimit-Reset' }
        '400': { description: Fichier manquant/invalide }
        '401': { description: Non authentifié (si requis) }
        '429':
          description: Limite atteinte
          headers:
            Retry-After: { description: Secondes avant un nouvel essai, schema: { type: integer, example: 60 } }
...